+++ src/scl/utility/hash/sdbm.h
@@ -10,6 +10,7 @@
 
 #include <cstdint>
 #include <ranges>
+#include <utility>
 
 #include "detail/base.h"
 
@@ -72,14 +73,12 @@ namespace scl::hash
      */
     template <::scl::hash::concepts::hashable_range Range>
     [[nodiscard]]
-    constexpr ::std::uint64_t sdbm(Range const & range, ::std::uint64_t h = 0ull)
+    constexpr ::std::uint64_t sdbm(Range && range, ::std::uint64_t h = 0ull)
     {
-        for (auto const c : range)
-        {
+        for (::std::ranges::range_value_t<Range> const c :
+            ::scl::hash::detail::read_only(::std::forward<Range>(range)))
             // cppcheck-suppress useStlAlgorithm
             h = ::scl::hash::detail::as_byte(c) + (h << 6) + (h << 16) - h;
-        }
-
         return h;
     }
 
@@ -93,9 +92,9 @@ namespace scl::hash
 
         template <::scl::hash::concepts::hashable_range Range>
         [[nodiscard]]
-        constexpr result_type operator()(Range const & range) const noexcept
+        constexpr result_type operator()(Range && range) const noexcept
         {
-            return ::scl::hash::sdbm(range);
+            return ::scl::hash::sdbm(::std::forward<Range>(range));
         }
     };
 
@@ -111,11 +110,12 @@ namespace scl::hash
  */
 
 /**
- * @fn scl::hash::sdbm_hasher::operator()(Range const & range) const
+ * @fn scl::hash::sdbm_hasher::operator()(Range && range) const
  * @brief Hashes @p range with @ref scl::hash::sdbm and its default seed.
  *
  * @tparam Range  Any type satisfying @ref scl::hash::concepts::hashable_range - a range of
- *                single trivially copyable bytes that is not a bounded array.
+ *                trivially copyable non-empty elements one byte wide that is not a bounded
+ *                array.
  * @param  range  Input range to hash.
  * @return 64-bit sdbm hash value of @p range.
  */